genericAssertion(): fix a few minor issues
The function stopped validating the 'l2address' assertion type properly
since the commit 039ef80. However, it is unlikely the bug had caused any
consequences because RackTables code asserts only the 'l2address0' type.

Fix that and put respective IAE to IRAE class conversion right to deliver
the message from l2addressForDatabase() as much intact as possible
(before this change genericAssertion() replaced the message and the
mention of IPoIB was not making it to the user interface).

Fix a similar conversion issue for the 'uint-vlan' and 'uint-vlan1'
types (this removes a lot of unnecessary text from the message). Ibid,
for clarity call assertStringArg() outside of the conversion block.