show object names in allocated units (Mantis#1711)
This change implements the original idea by Bram.

renderAtomGrid() generates forms built into rack diagrams to allow the
user to adjust rack design and to mark rack-specific problems and to
change rackspace allocation for an object. The function used to put a
checkbox into each atom of the diagram regardless if specific atoms
belong to the context (where they did not belong, the checkboxes were
disabled). This was supposedly done to keep the rack diagram unit rows
exactly the same height.

However, that was not convenient as the user could not correlate the
checkboxes of interest with neither names nor exact boundaries of known
objects in the rack and often had to rely on unit numbers only or to flip
between the default and the rackspace tabs.

Port some logic from renderRack() to renderAtomGrid() to show mounted
objects on the said forms. Amend addBulkSelectorJS() to keep the bulk
selector away from the object hyperlinks, which are much more useful if
they remain clickable, especially on the object rackspace tab.