renderServerConfigurationReport(): handle InnoDB failures better
The InnoDB test fails trying to create a trigger when binary logging is
enabled in the database server (in particular, that is required for the
old-style master-slave replication setup). This used to result in an
unhandled PDOException, which effectively blinded all other tests.

Use the exception to generate a separate specific error message and to
skip the InnoDB test only.