add and use RTTestCase::myStringStatic()
Even after the introduction of RTTestCase::myString() some tests
continued to use fixed strings for the database props naming in their
static setup methods. When such a test failed in a way that prevented the
teardown method to run, that would fail subsequent attempts to run the
setup method until the props were manually removed or the unit test
database contents was reset.

Introduce a static counterpart of the existing solution and update the
following classes to use it:

* EntityLinkTriggerTest
* GetChildrenListTest
* ObjectCircularReferenceTest
* ObjectLogTest