git-commit: say add/update in the default message
The script used to log "update <filename>" regardless if it was a new
file or a change to an existing file. To make it easier to follow the
history, have it log "add <filename>" in the former case. Replace a
long if() wrapper with an early return while at it.

[skip ci]